% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcaresearch.R
\name{care}
\alias{care}
\title{All care search}
\usage{
care(x, c1 = "Description", c2 = "Title")
}
\arguments{
\item{x}{is the data frame to be used}

\item{c1}{is the first column to be searched}

\item{c2}{is the second column to be searched}
}
\value{
A dataframe
}
\description{
All care search
}
