% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_dates.R
\name{collapse}
\alias{collapse}
\title{Collapse Dates}
\usage{
collapse(x, sup = "supplier_clean", .start = "start_y", .end = "end_y")
}
\arguments{
\item{x}{is the dataset to be collapsed}

\item{sup}{is the name of the supplier}

\item{.start}{is the variable containing the start year}

\item{.end}{is the variable containing the end year}
}
\description{
Collapse Dates
}
