% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_func.R
\name{con_format}
\alias{con_format}
\title{Function to Format Registries Ready to be Appended}
\usage{
con_format(
  x,
  tit = "Title",
  dep = "Department",
  desc = "Description",
  st = "Start.Date",
  en = "End.Date",
  rev = "Review.Date",
  frame = "Framework.Type",
  sup = "Supplier",
  council = "Council"
)
}
\arguments{
\item{x}{is the dataframe to be used}

\item{tit}{is the contract title variable}

\item{dep}{is the council department variable}

\item{desc}{is the description of the contract variable}

\item{st}{is the start date variable}

\item{en}{is the end date variable}

\item{rev}{is the review date variable}

\item{frame}{is the framework type variable}

\item{sup}{is the supplier name variable}

\item{council}{is the council variable}
}
\value{
a dataframe
}
\description{
Function to Format Registries Ready to be Appended
}
