% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractsearch.R
\name{home}
\alias{home}
\title{home care only search, intended to be piped after all-care search}
\usage{
home(x, c1 = "Title", c2 = "Description")
}
\arguments{
\item{x}{is the dataframe used}

\item{c1}{is the first column to be searched}

\item{c2}{is the second column to be searched}
}
\value{
A dataframe
}
\description{
home care only search, intended to be piped after all-care search
}
