% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resonly.R
\name{res}
\alias{res}
\title{search amongst care contracts for residential care}
\usage{
res(x, c1 = "Title", c2 = "Description")
}
\arguments{
\item{x}{is the dataframe input}

\item{c1}{is the first column to be searched}

\item{c2}{is the second column to be searched}
}
\value{
a dataframe
}
\description{
search amongst care contracts for residential care
}
